requirements.txt
astroid==2.4.2
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
peewee==3.13.3
pkg-resources==0.0.0
pylint==2.6.0
PyYAML==5.3.1
rope==0.17.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
